+++ results/verified/20260721T032358Z_DD-003_84238b76_2cbc13e66a/reproduce.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+set -eu
+exec make dd003-heterogeneous-sources